Rename support view to problem report
