+++ gui/src/renderer/components/ProblemReportStyles.tsx
