+++ gui/src/renderer/containers/SettingsPage.tsx
@@ -25,7 +25,7 @@ const mapDispatchToProps = (_dispatch: ReduxDispatch, props: IHistoryProps & IAp
     onClose: () => props.history.dismiss(),
     onViewSelectLanguage: () => props.history.push(RoutePath.selectLanguage),
     onViewAccount: () => props.history.push(RoutePath.accountSettings),
-    onViewSupport: () => props.history.push(RoutePath.support),
+    onViewSupport: () => props.history.push(RoutePath.problemReport),
     onViewPreferences: () => props.history.push(RoutePath.preferences),
     onViewAdvancedSettings: () => props.history.push(RoutePath.advancedSettings),
     onExternalLink: (url: string) => props.app.openUrl(url),
